Add color interpolation to Trinki method
For this change, I had to disect the code a bit to get it to work.

First off, I have to comment out the other rasterizers, as I want
to use float color and position values, and the other ones are
ill prepared for that.

Second, they do not output color interpolation yet, so thats why
they are removed.

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>